#18147 refactor for the integration test
